+++ src/exception.rs
@@ -8,7 +8,7 @@ use pattern::KLPair;
 
 /// A specialized hash map of pattern-score pairs.
 #[derive(Clone, Debug)]
-pub struct Exceptions(pub HashMap<String, Vec<u32>>);
+pub struct Exceptions(pub HashMap<String, Vec<u8>>);
 
 impl Exceptions {
     /// Creates an empty `Exceptions` map.
@@ -19,7 +19,7 @@ impl Exceptions {
     /// Inserts a Knuth-Liang exception pair into the map.
     ///
     /// If the pattern already existed, the old score is returned; if not, `None` is.
-    pub fn insert(&mut self, klpair: KLPair) -> Option<Vec<u32>> {
+    pub fn insert(&mut self, klpair: KLPair) -> Option<Vec<u8>> {
         let (p, score) = klpair;
         let Exceptions(ref mut m) = *self;
 
@@ -27,7 +27,7 @@ impl Exceptions {
     }
 
     /// Retrieves the score for each hyphenation point of `word`.
-    pub fn score(&self, word: &str) -> Option<&Vec<u32>> {
+    pub fn score(&self, word: &str) -> Option<&Vec<u8>> {
         let Exceptions(ref m) = *self;
         let w = word.to_lowercase();
